refactor/file-org: Move update_container_perms
Moves from ipc to access_container.rs
