+++ safe_authenticator/src/ipc.rs
@@ -9,7 +9,6 @@
 //! Inter-Process Communication utilities.
 
 use super::{AuthError, AuthFuture};
-use crate::access_container;
 use crate::app_auth::{app_state, AppState};
 use crate::client::AuthClient;
 use crate::config;
@@ -20,13 +19,12 @@ use futures::future::{self, Either};
 use futures::Future;
 use safe_core::ffi::ipc::resp::MetadataResponse as FfiUserMetadata;
 use safe_core::ipc::req::{
-    container_perms_into_permission_set, ContainerPermissions, IpcReq, ShareMDataReq,
+    IpcReq, ShareMDataReq,
 };
-use safe_core::ipc::resp::{AccessContainerEntry, IpcResp, UserMetadata, METADATA_KEY};
+use safe_core::ipc::resp::{IpcResp, UserMetadata, METADATA_KEY};
 use safe_core::ipc::{self, IpcError, IpcMsg};
-use safe_core::{mutable_data, Client, CoreError, FutureExt};
-use safe_nd::{Error as SndError, PublicKey, XorName};
-use std::collections::HashMap;
+use safe_core::{Client, CoreError, FutureExt};
+use safe_nd::{Error as SndError, XorName};
 use std::ffi::CString;
 
 /// Decodes a given encoded IPC message and returns either an `IpcMsg` struct or
@@ -102,55 +100,6 @@ pub fn decode_ipc_msg(
     }
 }
 
-/// Updates containers permissions (adds a given key to the permissions set)
-#[allow(clippy::implicit_hasher)]
-pub fn update_container_perms(
-    client: &AuthClient,
-    permissions: HashMap<String, ContainerPermissions>,
-    app_pk: PublicKey,
-) -> Box<AuthFuture<AccessContainerEntry>> {
-    let c2 = client.clone();
-
-    access_container::fetch_authenticator_entry(client)
-        .and_then(move |(_, mut root_containers)| {
-            let mut reqs = Vec::new();
-            let client = c2.clone();
-
-            for (container_key, access) in permissions {
-                let c2 = client.clone();
-                let mdata_info = fry!(root_containers
-                    .remove(&container_key)
-                    .ok_or_else(|| AuthError::NoSuchContainer(container_key.clone())));
-                let perm_set = container_perms_into_permission_set(&access);
-
-                let fut = client
-                    .get_mdata_version(*mdata_info.address())
-                    .and_then(move |version| {
-                        mutable_data::set_mdata_user_permissions(
-                            &c2,
-                            *mdata_info.address(),
-                            app_pk,
-                            perm_set,
-                            version + 1,
-                        )
-                        .map(move |_| (container_key, mdata_info, access))
-                    })
-                    .map_err(AuthError::from);
-
-                reqs.push(fut);
-            }
-
-            future::join_all(reqs).into_box()
-        })
-        .map(|perms| {
-            perms
-                .into_iter()
-                .map(|(container_key, dir, access)| (container_key, (dir, access)))
-                .collect()
-        })
-        .map_err(AuthError::from)
-        .into_box()
-}
 
 /// Encode `IpcMsg` into a `CString`, using base32 encoding.
 pub fn encode_response(msg: &IpcMsg) -> Result<CString, IpcError> {
